requirements.txt
langchain>=0.3.0
openai>=1.50.0
python-dotenv>=1.0.0
langchain-openai==0.3.33
pydantic>=2.11.10